opscore/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.127.0
google-generativeai==0.5.4
python-dotenv==1.0.1
itsdangerous==2.2.0
starlette==0.37.2
pydantic==2.7.1
pytest==8.2.0
httpx==0.27.0
